feat: update architecture diagram for OS2ID service
This commit updates the Mermaid diagram that represents the architecture of our OS2ID service. The changes include:

- Renaming of some components.
- The addition of a subgraph for the OS2ID service, which includes an 'ids' node and a 'User Management' node.
- 5 applications (Application 1-5) that interact with the OS2ID service.
- A Middleware component that interacts with the OS2ID service and a 'Legacy IDP' under 'Identiteter'.
- 3 'Identiteter' nodes, one of which is a legacy IDP that the Middleware component interacts with via a Legacy API.
- An 'Admin' node that interacts with the OS2ID service.

This updated diagram provides a clearer view of how our applications interact with the OS2ID service, how the OS2ID service communicates with the 'Identiteter' nodes, and how the Middleware component fits into this architecture.